day_7

+++ app/Http/Controllers/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\UserRequest;
+use App\Models\User;
 use App\Repositories\User\UserRepositoryInterface;
 use Illuminate\Http\Request;
 
@@ -24,47 +25,38 @@ class UserController extends Controller
     public function update(UserRequest $request, $id)
     {
         $data = $request->validated();
+        $user = $this->userRepository->show($id);
 
         if ($request->hasFile('image')) {
             $image = $request->file('image');
             $imageName = time() . '_' . $image->getClientOriginalName();
 
             // Delete old image if exists
-            $user = $this->userRepository->show($id);
             if ($user->image && file_exists(public_path('images/' . $user->image))) {
                 unlink(public_path('images/' . $user->image));
             }
 
-            // Store new image
             $image->move(public_path('images'), $imageName);
             $data['image'] = $imageName;
         }
 
-        $user = $this->userRepository->update($data, $id);
-
-        if ($request->ajax()) {
-            return response()->json([
-                'success' => true,
-                'message' => 'Profile updated successfully',
-                'user' => $user
-            ]);
-        }
-
-        return redirect()->route('user.show', $id)->with('success', 'Profile updated successfully');
+        $this->userRepository->update($data, $id);
+        return redirect()->route('users.index')->with('success', 'User updated successfully');
     }
 
     public function store(UserRequest $request)
     {
-        $lastUser = $this->userRepository->store($request->all());
-
+        $data = $request->validated();
+        
         if ($request->hasFile('image')) {
             $image = $request->file('image');
             $imageName = time() . '_' . $image->getClientOriginalName();
             $image->move(public_path('images'), $imageName);
-            $this->userRepository->update(['image' => $imageName], $lastUser->id);
+            $data['image'] = $imageName;
         }
 
-        return redirect()->route('users.index');
+        $this->userRepository->store($data);
+        return redirect()->route('users.index')->with('success', 'User created successfully');
     }
 
     public function create()
@@ -90,26 +82,56 @@ class UserController extends Controller
         return view('users.edit', compact('user'));
     }
 
+    // public function updateAnotherUser(UserRequest $request, $id)
+    // {
+    //     $data = $request->validated();
+
+    //     if ($request->hasFile('image')) {
+    //         $image = $request->file('image');
+    //         $imageName = time() . '_' . $image->getClientOriginalName();
+
+    //         // Delete old image if exists
+    //         $user = $this->userRepository->show($id);
+    //         if ($user->image && file_exists(public_path('images/' . $user->image))) {
+    //             unlink(public_path('images/' . $user->image));
+    //         }
+
+    //         // Store new image
+    //         $image->move(public_path('images'), $imageName);
+    //         $data['image'] = $imageName;
+    //     }
+
+    //     $this->userRepository->updateAnotherUser($data, $id);
+    //     return redirect()->route('users.index')->with('success', 'User updated successfully');
+    // }
+
     public function updateAnotherUser(UserRequest $request, $id)
     {
         $data = $request->validated();
 
-        if ($request->hasFile('image')) {
+        if ($request->hasFile('image') && $request->file('image')->isValid()) {
             $image = $request->file('image');
             $imageName = time() . '_' . $image->getClientOriginalName();
 
-            // Delete old image if exists
-            $user = $this->userRepository->show($id);
+            // Ensure directory exists
+            if (!file_exists(public_path('images'))) {
+                mkdir(public_path('images'), 0755, true);
+            }
+
+            // Delete old image if it exists
+            $user = User::findOrFail($id);
             if ($user->image && file_exists(public_path('images/' . $user->image))) {
                 unlink(public_path('images/' . $user->image));
             }
 
-            // Store new image
+            // Save new image
             $image->move(public_path('images'), $imageName);
             $data['image'] = $imageName;
         }
 
-        $this->userRepository->updateAnotherUser($data, $id);
+        // Update user
+        User::findOrFail($id)->update($data);
+
         return redirect()->route('users.index')->with('success', 'User updated successfully');
     }
 }
